refactor: Rename SavedTasks to Storage, update functionality to only interacting with persistent storage
